requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
Django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-ninja==1.4.1
dotenv==0.9.9
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
